requirements.txt
fpdf2
GitPython
inquirer
pathspec
pytest